# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TradePilot API
  description: REST API for TradePilot crypto trading bot - control, monitoring, backtesting.
  version: 1.0.0

servers:
  - url: /
    description: Relative (e.g. https://web-production-f8308.up.railway.app)

paths:
  /api/auth/signin:
    post:
      summary: Sign in
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: JWT token }
        '401': { description: Invalid credentials }

  /api/auth/verify:
    get:
      summary: Verify JWT token
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Token valid }
        '401': { description: Invalid or expired token }

  /api/status:
    get:
      summary: Bot status
      tags: [Status]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Bot status, uptime, mode }

  /api/trades:
    get:
      summary: List recent trades
      tags: [Trades]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
      responses:
        '200': { description: List of trades }

  /api/trades/{trade_id}:
    get:
      summary: Get trade details
      tags: [Trades]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: trade_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Trade details }
        '400': { description: Invalid trade ID }
        '404': { description: Trade not found }

  /api/backtest/run:
    post:
      summary: Run backtest
      tags: [Backtest]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pair: { type: string, default: 'BTC-USD' }
                days: { type: integer, default: 30, minimum: 1, maximum: 90 }
                initial_balance: { type: number, default: 100000, minimum: 100, maximum: 10000000 }
      responses:
        '200': { description: Backtest results }
        '500': { description: Backtest failed }

  /api/backtest/list:
    get:
      summary: List backtests
      tags: [Backtest]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200': { description: List of backtests }

  /api/settings:
    get:
      summary: Get settings
      tags: [Settings]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current settings }
    post:
      summary: Update settings
      tags: [Settings]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ema_period: { type: integer, minimum: 5, maximum: 200 }
                rsi_period: { type: integer, minimum: 5, maximum: 50 }
                volume_multiplier: { type: number, minimum: 0.1, maximum: 10 }
                min_confidence: { type: number, minimum: 0, maximum: 100 }
      responses:
        '200': { description: Settings updated }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT